feat: api documentation ( DAC-388 )
