+++ dapps-certification-interface/dapps-certification-interface.cabal
@@ -6,6 +6,6 @@ maintainer:         shea.levy@iohk.io
 
 library
     exposed-modules:  IOHK.Certification.Interface
-    build-depends:    base, aeson, text
+    build-depends:    base, aeson, text, swagger2
     hs-source-dirs:   src
     default-language: Haskell2010
