+++ dapps-certification-persistence/dapps-certification-persistence.cabal
@@ -16,5 +16,7 @@ library
                  , text
                  , time
                  , aeson
+                 , swagger2
+                 , lens
     hs-source-dirs:   src
     default-language: Haskell2010
