+++ nix/materialized/x86_64-linux/.plan.nix/dapps-certification-persistence.nix
@@ -44,6 +44,8 @@
           (hsPkgs."text" or (errorHandler.buildDepError "text"))
           (hsPkgs."time" or (errorHandler.buildDepError "time"))
           (hsPkgs."aeson" or (errorHandler.buildDepError "aeson"))
+          (hsPkgs."swagger2" or (errorHandler.buildDepError "swagger2"))
+          (hsPkgs."lens" or (errorHandler.buildDepError "lens"))
           ];
         buildable = true;
         modules = [
