+++ plutus-certification.cabal
@@ -41,6 +41,10 @@ library
     eventuo11y ^>= 0.3.2,
     eventuo11y-batteries ^>= 0.2.1,
     time ^>= 1.11.1.1,
+    servant-swagger ^>= 1.1.11,
+    swagger2,
+    servant-swagger-ui,
+    lens,
     filepath,
     temporary,
     dapps-certification-persistence,
@@ -55,7 +59,10 @@ library
     Plutus.Certification.Server
     Plutus.Certification.Local
     Plutus.Certification.GithubClient
-  other-modules: Paths_plutus_certification
+  other-modules:
+    Paths_plutus_certification
+    Plutus.Certification.API.Routes
+    Plutus.Certification.API.Swagger
   autogen-modules: Paths_plutus_certification
   default-language: Haskell2010
 
@@ -80,6 +87,8 @@ executable plutus-certification
     bytestring ^>= 0.11.1.0,
     plutus-certification,
     dapps-certification-persistence,
+    swagger2,
+    servant-swagger-ui,
   main-is: Main.hs
   hs-source-dirs: server
   other-modules: Paths_plutus_certification
